sc2: add config to support multi-ddr firmware [1/3]
PD#SWPL-40605

Problem:
SC2 only support 1 ddr firmware

Solution:
Add config to support multi-ddr firmware

Verify:
ah212

+++ cmd/cmd_startdsp.c
@@ -82,8 +82,8 @@ void set_dsp_clk(uint32_t id, uint32_t freq_sel)
 
 static int do_startdsp(cmd_tbl_t *cmdtp, int flag, int argc, char * const argv[])
 {
-	unsigned long addr;
-	unsigned int dspid;
+	uint32_t addr;
+	uint32_t dspid;
 	uint32_t freq_sel;
 	int ret=0;
 	if (argc <= 1) {
